feat(sliding-sync): add e2ee extension types
