+++ crates/ruma-client-api/src/sync/sync_events.rs
@@ -1,6 +1,7 @@
 //! `GET /_matrix/client/*/sync`
 
 use js_int::UInt;
+use ruma_common::OwnedUserId;
 use serde::{self, Deserialize, Serialize};
 
 pub mod v3;
@@ -30,5 +31,33 @@ impl UnreadNotificationsCount {
     /// Returns true if there are no notification count updates.
     pub fn is_empty(&self) -> bool {
         self.highlight_count.is_none() && self.notification_count.is_none()
+    }
+}
+
+
+/// Information on E2E device updates.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
+pub struct DeviceLists {
+    /// List of users who have updated their device identity keys or who now
+    /// share an encrypted room with the client since the previous sync
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub changed: Vec<OwnedUserId>,
+
+    /// List of users who no longer share encrypted rooms since the previous sync
+    /// response.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub left: Vec<OwnedUserId>,
+}
+
+impl DeviceLists {
+    /// Creates an empty `DeviceLists`.
+    pub fn new() -> Self {
+        Default::default()
+    }
+
+    /// Returns true if there are no device list updates.
+    pub fn is_empty(&self) -> bool {
+        self.changed.is_empty() && self.left.is_empty()
     }
 }
